Expose all the available isolation level constants.
Note that what used to be SERIALIZABLE in previous PostgreSQL versions is
currently REPEATABLE READ, and on older PostgreSQL versions the two are the
same. So the default has changed to REPEATABLE READ because anywhere the
difference matters, the semantics you already know is that one. Please read
the PostgreSQL documentation for further info.