T179375 improve failover behaviour (#1)
* Check if getconn() still works during abort()

During failover, a transaction abort() may be issued by Zope because a 
query failed when the database was shut down.

This means that getconn() during abort() will also fail. This case was 
not covered yet.

* Add new NotSupportedError to connection_errors, fix bug in _abort()

When using replicas, we get the occasional read-write error because 
we're connected to the wrong database during failover.

This error belongs in the class of errors which we want to recover from 
via retry.